HV:dead code cleaning
In FuSa scope, we have following Application Constraints:
The INVEPT type single-context invalidation
shall be available on the physical platform.
So remove the related dead code.

As we have the application constraint:
"The physical platform shall guarantee that the following CPUID leaves
are supported: 0H, 1H, 7H, 80000000H, 80000001H, 80000007H, and 80000008H",
so related 'if' check can be eliminated.

Signed-off-by: liujunming <junming.liu@intel.com>