treewide: add missing declaration of acrn_vcpu or acrn_vm
Some headers use acrn_vm and/or acrn_vcpu structure types without including
their definitions in vm.h and/or vcpu.h. This result in some warnings when
compiling the headers themselves for testing purposes.

This patch adds the missing includes to make sure that the structures are
properly defined before they are used.

This cannot be done to vcpuid.h and virtual_cr.h, however, as vm.h and
vcpu.h relies on these two to define the structures. As a result a
declaration of the structures are added instead.

Signed-off-by: Junjie Mao <junjie.mao@intel.com>