cpu: fix a potential dead loop in start_pcpus
Assuming current CPU ID is 'x' and bit 'x' is set in parameter 'mask',
it leads to a dead loop in start_pcpus with current implementation.

Fix this issue by getting the ID of the next to-be-started CPU
before jummping to next iteration when the corresponding bit of
the current CPU ID is set in in parameter 'mask'.

Signed-off-by: Shiqing Gao <shiqing.gao@intel.com>