vm: protect VM state transitions with spinlock
The current VM state transitions are not protected by locks, which
could lead to unexpected behavior when multiple cores attempt to
operate the same VM at the same time.

Similar to what has been applied the ACRN mainline, this patch
introduces the per-VM state lock (named vm_lock in compliance with
SAS), use the lock to protect calls to VM lifecycle operations and
clarify the pre-conditions on the VM states that these functions
assume.

Tracked-On: ACCB-78
Signed-off-by: Junjie Mao <junjie.mao@intel.com>