CCB-169: fix the boot module
The 64-bit starting code was unnecessary as the hypervisor never
expected itself to be started from long mode.

The save context behavior in 32/64 bit mode is unnecessary.

In current code, the initial host CR0 value is fixed with bit or operation
which depends on the previous CR0 value set by BIOS. This operation will
cause potential issue if BIOS set bit CD to 1 to disable cache.
Modify the operation to set the host CR0 a fixed value 0x80000033H to
avoid the dependency and potential issue.

This patch is used to fix these issues.

Signed-off-by: Junjun Shan <junjun.shan@intel.com>