Refine the emulation of PCI bar register
When writing "val" to bar, just update the bar register
to ("val" & bar->mask) | bar->fixed.
This method also applies to ~0, which is used to get
PCI bar size.

If the new base address is in the PCI hole,
needs do the following operations:
(1) update bar->base to the new base address
(2) remap the BAR space between the new GPA of BAR base and its HPA.

If the new base address is out of PCI hole,
just update bar->base to 0, which means its a invalid GPA.

Tracked-On: ACCB-172

Signed-off-by: liujunming <junming.liu@intel.com>