HV:hot fix bugs in vperipheral module
According to design, the virtual command register [bit 10] in PCI
configuration space should be set as 1H to disable INTx
interrupt.
The virtual MSI data register [bit 16 ... bit 32] should be
read-only bits.

But in the current code, the virtual command register [bit 10]
is not be set as 1H, the virtual MSI data register [bit 16 ... bit 32]
is implemented as write-read bits.

Signed-off-by: Xiangyang Wu <xiangyang.wu@intel.com>