HV:fix bugs found during software failure analysis
Unintended NMI will be injected by hypervisor when a NMI is triggered
by hardware in non root mode of safety VM if the descriptor of vector 2
in guest IDT is a task gate.
About similar case in the native, only exception is handled and this NMI
will be discarded.
In this case, the task switch will cause a VM exit during event delivery,
the current hypervisor will record the NMI event in the event pending
field of vCPU, will inject #GP to guest first and inject NMI to guest
after the next VM exit.
This patch remove the NMI injection logic in this case.
Refer to FMEA-303028 and FMEA-303029 for more detail.

Miss delay between the INIT IPI and first startup IPI to start
physical APs.
Refer to FMEA-277676 for more detail.

This patch fixs these two bugs.

Signed-off-by: Xiangyang Wu <xiangyang.wu@intel.com>